Ibp 3669 prepare planting

+++ src/main/java/org/ibp/api/java/impl/middleware/design/ExperimentalDesignServiceImpl.java
@@ -3,8 +3,11 @@ package org.ibp.api.java.impl.middleware.design;
 import org.generationcp.middleware.domain.dms.ExperimentDesignType;
 import org.generationcp.middleware.domain.etl.MeasurementVariable;
 import org.generationcp.middleware.manager.api.WorkbenchDataManager;
+import org.generationcp.middleware.pojos.ims.TransactionStatus;
 import org.generationcp.middleware.pojos.workbench.CropType;
 import org.generationcp.middleware.service.api.study.StudyGermplasmDto;
+import org.generationcp.middleware.service.impl.inventory.PlantingServiceImpl;
+import org.ibp.api.exception.ApiRequestValidationException;
 import org.ibp.api.exception.ForbiddenException;
 import org.ibp.api.java.design.DesignLicenseService;
 import org.ibp.api.java.design.ExperimentalDesignService;
@@ -22,7 +25,9 @@ import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.validation.BindingResult;
 import org.springframework.validation.FieldError;
+import org.springframework.validation.MapBindingResult;
 
 import javax.annotation.Resource;
 import java.util.ArrayList;
@@ -66,6 +71,9 @@ public class ExperimentalDesignServiceImpl implements ExperimentalDesignService 
 	@Resource
 	private DesignLicenseService designLicenseService;
 
+	@Resource
+	private PlantingServiceImpl plantingService;
+
 	@Override
 	public void generateAndSaveDesign(final String cropName, final int studyId, final ExperimentalDesignInput experimentalDesignInput) {
 		this.studyValidator.validate(studyId, true);
@@ -104,7 +112,22 @@ public class ExperimentalDesignServiceImpl implements ExperimentalDesignService 
 	public void deleteDesign(final int studyId) {
 		this.studyValidator.validate(studyId, true, true);
 		this.experimentalDesignValidator.validateExperimentalDesignExistence(studyId, true);
-		this.experimentDesignMiddlewareService.deleteStudyExperimentDesign(studyId);
+		//FIXME This service should be validating at least the NO existence of sub-obs. This validation is delegated to the frontend
+		final BindingResult errors = new MapBindingResult(new HashMap<String, String>(), Integer.class.getName());
+		final Integer pendingTransactions =
+			this.plantingService.getPlantingTransactionsByStudyId(studyId, TransactionStatus.PENDING).size();
+		final Integer confirmedTransactions =
+			this.plantingService.getPlantingTransactionsByStudyId(studyId, TransactionStatus.CONFIRMED).size();
+		if (pendingTransactions > 0 || confirmedTransactions > 0) {
+			errors.reject("study.has.pending.or.confirmed.transactions");
+			throw new ApiRequestValidationException(errors.getAllErrors());
+		}
+		try {
+			this.experimentDesignMiddlewareService.deleteStudyExperimentDesign(studyId);
+		} catch (final Exception e) {
+			errors.reject("experimental.design.general.error");
+			throw new ApiRequestValidationException(errors.getAllErrors());
+		}
 	}
 
 	@Override
